Rename parser.cpp to main.cpp, and new-parser to parser.
