+++ compiler/AMBuilder
@@ -52,8 +52,8 @@ binary.sources += [
   'expressions.cpp',
   'lexer.cpp',
   'libpawnc.cpp',
+  'main.cpp',
   'name-resolution.cpp',
-  'new-parser.cpp',
   'output-buffer.cpp',
   'parser.cpp',
   'pawncc.cpp',
