+++ compiler/lexer.cpp
@@ -51,8 +51,8 @@
 #include "lexer.h"
 #include "lexer-inl.h"
 #include "libpawnc.h"
-#include "new-parser.h"
 #include "output-buffer.h"
+#include "parser.h"
 #include "sc.h"
 #include "sci18n.h"
 #include "sclist.h"
