+++ compiler/name-resolution.cpp
@@ -25,8 +25,8 @@
 #include "array-helpers.h"
 #include "errors.h"
 #include "expressions.h"
-#include "new-parser.h"
 #include "parse-node.h"
+#include "parser.h"
 #include "sc.h"
 #include "sclist.h"
 #include "sctracker.h"
